requirements.txt
flask
openai
jinja2
mysql-connector-python
psycopg2-binary
requests         
